+++ src/main/groovy/com/episode6/hackit/deployable/DeployableKotlinJarPlugin.groovy
@@ -34,11 +34,11 @@ class DeployableKotlinJarPlugin implements Plugin<Project> {
       classifier = 'sources'
     }
 
-    project.deployable {
-      mainArtifact {
+    project.deployable.publication {
+      main {
         artifact project.jar
       }
-      publication {
+      amend {
         artifact project.javadocJar
         artifact project.sourcesJar
       }